ID を ファイルネームを使用するように変更

+++ Infinity/Core/ShopState.cs
@@ -125,7 +125,7 @@ public partial class ShopState : Node
 
         // ToDo: 外部化していいかも
         // プレイヤーのお金を減らす
-        Main.PlayerState.AddEffect(new MoneyEffect { Value = -minion.Price });
+        Main.PlayerState.AddEffect(new MoneyEffect { Value = -(int)minion.Price });
         Main.PlayerState.SolveEffect();
 
         // Player がすでに Minion を所持していたらレベルを上げる
@@ -225,7 +225,7 @@ public partial class ShopState : Node
 
             // プレイヤーのお金を増やす
             // TODO: 売却価格を売値と同じにしています
-            Main.PlayerState.AddEffect(new MoneyEffect { Value = minion.Price });
+            Main.PlayerState.AddEffect(new MoneyEffect { Value = (int)minion.Price });
             Main.PlayerState.SolveEffect();
         }
     }
